pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
dependencies = [
    "flask",
    "gunicorn",
    "python-dotenv",
    "flake8",
    "psycopg>=3.3.2",
    "validators>=0.35.0",
]
